+++ src/components/@common/avartar/index.stories.ts
@@ -1,5 +1,5 @@
-import { Avatar } from './avartar';
 import type { Meta, StoryObj } from '@storybook/react';
+import { Avatar } from '.';
 
 const meta: Meta<typeof Avatar> = {
   title: 'Design System/Atoms/Avatar',
