+++ src/components/@common/button/index.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Button from './button';
+import Button from '.';
 
 const meta: Meta<typeof Button> = {
   title: 'Design System/Atoms/Button',
