+++ src/components/@common/logo/index.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Logo from './logo';
+import Logo from '.';
 
 const meta: Meta<typeof Logo> = {
   title: 'Design System/Atoms/Logo',
