--- src/components/@common/polarChart/polarChart.stories.ts
@@ -1,17 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import { PolarChart } from './polarChart';
-
-const meta: Meta<typeof PolarChart> = {
-  title: 'DESIGN SYSTEM/Atoms/PolarChart',
-  component: PolarChart,
-  tags: ['autodocs'],
-};
-
-export default meta;
-type Story = StoryObj<typeof PolarChart>;
-
-export const Primary: Story = {
-  args: {
-    data: [],
-  },
-};
