+++ src/components/@common/questionProgress/index.tsx
@@ -1,7 +1,7 @@
 import { ProgressStatus } from '@/types';
 import React from 'react';
-import { Icon } from '../icon/icon';
-import { Text, TextColor } from '../text/text';
+import { Icon } from '../icon';
+import { Text, TextColor } from '../text';
 
 interface IQuestionProgress {
   progress: ProgressStatus;
