+++ src/components/@common/text/index.tsx
